 \name{algae.sols}
          \docType{data}
          \alias{algae.sols}
          \title{The solutions for the test data set for predicting algae blooms}
          \description{
            This data set contains the values of the 7 harmful algae for
          the 140 test observations in the test set \code{test.algae}.  
          }
          \usage{algae.sols}
          \format{A data frame with 140 observations and 7 columns.}
          \source{ERUDIT \url{http://www.erudit.de/} - European Network for Fuzzy Logic and
          Uncertainty Modelling in Information Technology. }
          \keyword{datasets}
