\name{DMwR2-package}
\alias{DMwR2-package}
\alias{DMwR2}
\docType{package}
\title{
Functions and data for the second edition of the book "Data Mining with R"
}
\description{
This package includes functions and data accompanying the book
"Data Mining with R, learning with case studies - 2nd Edition" by Luis Torgo,
published by CRC Press
}
\author{
Luis Torgo

Maintainer: Luis Torgo \email{ltorgo@dcc.fc.up.pt}
}
\references{
Torgo, L. (2016) \emph{Data Mining using R: learning with case studies -
2nd edition},
CRC Press . \url{http://www.dcc.fc.up.pt/~ltorgo/DataMiningWithR}  
}
\keyword{ package }
