 \name{GSPC}
          \docType{data}
          \alias{GSPC}
          \title{A set of daily quotes for SP500}
          \description{
	    This is a \code{xts} object containing the daily quotes of
          the SP500 sotck index from 1970-01-02 till 2009-09-15 (10,022
          daily sessions). For each day information is given on the
          Open, High, Low and Close prices, and also for the Volume and
          Adjusted close price.   
          }
          \usage{GSPC}
          \format{A \code{xts} object with a data matrix with 10,022
          rows and 6 columns.}
          \source{Yahoo Finance \url{http://finance.yahoo.com/}  }
          \keyword{datasets}
