 \name{sales}
          \docType{data}
          \alias{sales}
          \title{A data set with sale transaction reports}
          \description{
	    This data frame contains 401,146 transaction reports. Each
          report is made by a salesperson identified by an ID and
          reports the quantity sold of some product. The data set
          caontins information on 5 variables: ID (salesperson ID), Prod
          (product ID), Quant (the sold quantity), Val (the reported
          value of the transaction) and Insp (a factor containing
          information on a inspection of the report with possible values
          'ok',' fraud' or 'unkn').
          }
          \usage{sales}
          \format{A data frame with 401,146 rows and 5 columns}
          \source{Undisclosed}
          \keyword{datasets}
