 \name{sp500}
          \docType{data}
          \alias{sp500}
          \title{A set of daily quotes for SP500 in CSV Format}
          \description{
	    This is a \code{CSV} file containing the daily quotes of
          the SP500 sotck index from 1970-01-02 till 2015-12-25 . For
          each day information is given on the Open, High, Low and Close
          prices, and also for the Volume and  Adjusted close price.   
          }
          \usage{sp500}
          \format{A \code{CSV} file with a data matrix.}
          \source{Yahoo Finance \url{http://finance.yahoo.com/}  }
          \keyword{datasets}
