 \name{test.algae}
          \docType{data}
          \alias{test.algae}
          \title{Testing data for predicting algae blooms}
          \description{
            This data set contains observations on 11 variables as well
          as the concentration levels of 7 harmful algae. Values were
          measured in several European rivers. The 11 predictor
          variables include 3 contextual variables (season, size and
          speed) describing the water sample, plus 8 chemical
          concentration measurements.  
          }
          \usage{test.algae}
          \format{A data frame with 140 observations and 18 columns.}
          \source{ERUDIT \url{http://www.erudit.de/} - European Network for Fuzzy Logic and
          Uncertainty Modelling in Information Technology. }
          \keyword{datasets}
